change destroy case and specs

+++ app/cases/contact_list/destroy.rb
@@ -1,13 +1,13 @@
 module ContactList
   class Destroy < Micro::Case
-    attributes :contact
+    attributes :contact, :user
 
     def call!
-      if destroy_contact
-        Success result: { message: "Contato deletado com sucesso!" }
-      else
-        Failure result: { message: "Não foi possivel deletar o contato!" }
-      end
+      return Failure result: { message: "Voce não pode deletar este contato!" } unless check_user
+
+      destroy_contact
+
+      Success result: { message: "Contato deletado com sucesso!" }
     rescue => e
       Failure result: { message: "Não foi possivel deletar o contato!", error: e }
     end
@@ -15,8 +15,14 @@ module ContactList
     private
 
     def destroy_contact
-      contact.address&.destroy!
-      contact.destroy!
+      if contact.user_id == user.id
+        contact.address&.destroy!
+        contact.destroy!
+      end
+    end
+
+    def check_user
+      contact.user_id == user.id
     end
   end
 end
